+++ config/mapprovider/yamlmapprovider/deprecated.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//       http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,15 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package config // import "go.opentelemetry.io/collector/config"
+package yamlmapprovider // import "go.opentelemetry.io/collector/config/mapprovider/yamlmapprovider"
 
 import (
-	"context"
+	"go.opentelemetry.io/collector/confmap/provider/yamlprovider"
 )
 
-// MapConverter is a converter interface for the config.Map that allows distributions
-// (in the future components as well) to build backwards compatible config converters.
-type MapConverter interface {
-	// Convert applies the conversion logic to the given "cfgMap".
-	Convert(ctx context.Context, cfgMap *Map) error
-}
+// Deprecated: [v0.53.0] use yamlprovider.New
+var New = yamlprovider.New