+++ confmap/converter/overwritepropertiesconverter/properties_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package overwritepropertiesmapconverter
+package overwritepropertiesconverter
 
 import (
 	"context"
@@ -21,14 +21,14 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/confmap"
 )
 
 func TestOverwritePropertiesConverter_Empty(t *testing.T) {
 	pmp := New(nil)
-	cfgMap := config.NewMapFromStringMap(map[string]interface{}{"foo": "bar"})
-	assert.NoError(t, pmp.Convert(context.Background(), cfgMap))
-	assert.Equal(t, map[string]interface{}{"foo": "bar"}, cfgMap.ToStringMap())
+	conf := confmap.NewFromStringMap(map[string]interface{}{"foo": "bar"})
+	assert.NoError(t, pmp.Convert(context.Background(), conf))
+	assert.Equal(t, map[string]interface{}{"foo": "bar"}, conf.ToStringMap())
 }
 
 func TestOverwritePropertiesConverter(t *testing.T) {
@@ -40,18 +40,18 @@ func TestOverwritePropertiesConverter(t *testing.T) {
 	}
 
 	pmp := New(props)
-	cfgMap := config.NewMap()
-	require.NoError(t, pmp.Convert(context.Background(), cfgMap))
-	keys := cfgMap.AllKeys()
+	conf := confmap.New()
+	require.NoError(t, pmp.Convert(context.Background(), conf))
+	keys := conf.AllKeys()
 	assert.Len(t, keys, 4)
-	assert.Equal(t, "2s", cfgMap.Get("processors::batch::timeout"))
-	assert.Equal(t, "3s", cfgMap.Get("processors::batch/foo::timeout"))
-	assert.Equal(t, "foo:9200,foo2:9200", cfgMap.Get("exporters::kafka::brokers"))
-	assert.Equal(t, "localhost:1818", cfgMap.Get("receivers::otlp::protocols::grpc::endpoint"))
+	assert.Equal(t, "2s", conf.Get("processors::batch::timeout"))
+	assert.Equal(t, "3s", conf.Get("processors::batch/foo::timeout"))
+	assert.Equal(t, "foo:9200,foo2:9200", conf.Get("exporters::kafka::brokers"))
+	assert.Equal(t, "localhost:1818", conf.Get("receivers::otlp::protocols::grpc::endpoint"))
 }
 
 func TestOverwritePropertiesConverter_InvalidProperty(t *testing.T) {
 	pmp := New([]string{"=2s"})
-	cfgMap := config.NewMap()
-	assert.Error(t, pmp.Convert(context.Background(), cfgMap))
+	conf := confmap.New()
+	assert.Error(t, pmp.Convert(context.Background(), conf))
 }