+++ confmap/provider/fileprovider/mapprovider.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package filemapprovider // import "go.opentelemetry.io/collector/config/mapprovider/filemapprovider"
+package fileprovider // import "go.opentelemetry.io/collector/confmap/provider/fileprovider"
 
 import (
 	"context"
@@ -23,14 +23,14 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/confmap"
 )
 
 const schemeName = "file"
 
-type mapProvider struct{}
+type provider struct{}
 
-// New returns a new config.MapProvider that reads the configuration from a file.
+// New returns a new confmap.Provider that reads the configuration from a file.
 //
 // This Provider supports "file" scheme, and can be called with a "uri" that follows:
 //   file-uri		= "file:" local-path
@@ -43,33 +43,33 @@ type mapProvider struct{}
 // `file:/path/to/file` - absolute path (unix, windows)
 // `file:c:/path/to/file` - absolute path including drive-letter (windows)
 // `file:c:\path\to\file` - absolute path including drive-letter (windows)
-func New() config.MapProvider {
-	return &mapProvider{}
+func New() confmap.Provider {
+	return &provider{}
 }
 
-func (fmp *mapProvider) Retrieve(_ context.Context, uri string, _ config.WatcherFunc) (config.Retrieved, error) {
+func (fmp *provider) Retrieve(_ context.Context, uri string, _ confmap.WatcherFunc) (confmap.Retrieved, error) {
 	if !strings.HasPrefix(uri, schemeName+":") {
-		return config.Retrieved{}, fmt.Errorf("%v uri is not supported by %v provider", uri, schemeName)
+		return confmap.Retrieved{}, fmt.Errorf("%v uri is not supported by %v provider", uri, schemeName)
 	}
 
 	// Clean the path before using it.
 	content, err := ioutil.ReadFile(filepath.Clean(uri[len(schemeName)+1:]))
 	if err != nil {
-		return config.Retrieved{}, fmt.Errorf("unable to read the file %v: %w", uri, err)
+		return confmap.Retrieved{}, fmt.Errorf("unable to read the file %v: %w", uri, err)
 	}
 
 	var data map[string]interface{}
 	if err = yaml.Unmarshal(content, &data); err != nil {
-		return config.Retrieved{}, fmt.Errorf("unable to parse yaml: %w", err)
+		return confmap.Retrieved{}, fmt.Errorf("unable to parse yaml: %w", err)
 	}
 
-	return config.NewRetrievedFromMap(config.NewMapFromStringMap(data)), nil
+	return confmap.NewRetrievedFromMap(confmap.NewFromStringMap(data)), nil
 }
 
-func (*mapProvider) Scheme() string {
+func (*provider) Scheme() string {
 	return schemeName
 }
 
-func (*mapProvider) Shutdown(context.Context) error {
+func (*provider) Shutdown(context.Context) error {
 	return nil
 }